Some refactoring"
